# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Teamwork is an internal social network for employees of an organization. The goal of this application is to facilitate more interaction between colleagues and promote team bonding",
    "version": "1.0.0",
    "title": "Teamwork Api",
    "contact": {
      "email": "odinakajoy909@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "api-teamwork-by-joy.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/categories": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "creates category",
        "description": "Adds an category to the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "create category",
            "description": "category details to add",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Category successfully created"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": [
          "category"
        ],
        "summary": "displays all categories",
        "description": "displays all categories in the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Category found"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/articles": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "creates article",
        "description": "Adds an article to the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "create article",
            "description": "article details to add",
            "schema": {
              "$ref": "#/definitions/Article"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Article successfully posted"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/gifs": {
      "post": {
        "tags": [
          "gif"
        ],
        "summary": "creates gif",
        "description": "Adds an gif to the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "create gif",
            "description": "gif details to add",
            "schema": {
              "$ref": "#/definitions/Gif"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Gif image successfully posted"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/flags": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "display flags",
        "description": "display flags in the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Flags found"
          },
          "404": {
            "description": "No flag found"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "display all articles and gifs",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "success"
          },
          "404": {
            "description": "No article and gif found"
          },
          "422": {
            "description": "Invalid input, object invalid"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "required": [
        "category"
      ],
      "properties": {
        "category": {
          "type": "string"
        }
      }
    },
    "Article": {
      "type": "object",
      "required": [
        "categoryId",
        "title",
        "article",
        "articleImage",
        "userId"
      ],
      "properties": {
        "categoryId": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "article": {
          "type": "string"
        },
        "articleImage": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        }
      }
    },
    "Gif": {
      "type": "object",
      "required": [
        "title",
        "gif",
        "userId"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "gif": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        }
      }
    },
...
  }
}